[WEB-9752] Move sitemap filter out of lib/api into lib/sitemap
